+++ src/python/twitter/pants/tasks/scala_compile.py
@@ -18,28 +18,23 @@ __author__ = 'John Sirois'
 
 import os
 import shutil
-import textwrap
 
 from collections import defaultdict
-from contextlib import closing
-from xml.etree import ElementTree
 
-from twitter.common.collections import OrderedDict
-from twitter.common.contextutil import open_zip as open_jar, temporary_dir, temporary_file_path
-from twitter.common.dirutil import safe_mkdir, safe_open, safe_rmtree
+from twitter.common.contextutil import  temporary_dir
+from twitter.common.dirutil import safe_mkdir, safe_rmtree
 
-from twitter.pants import get_buildroot, is_scala, is_scalac_plugin
+from twitter.pants import  is_scalac_plugin
 from twitter.pants.base.target import Target
 from twitter.pants.targets.scala_library import ScalaLibrary
 from twitter.pants.targets.scala_tests import ScalaTests
 from twitter.pants.targets import resolve_target_sources
 from twitter.pants.tasks import TaskError
-from twitter.pants.tasks.binary_utils import find_java_home, nailgun_profile_classpath
 from twitter.pants.tasks.jvm_compiler_dependencies import Dependencies
 from twitter.pants.tasks.jvm_dependency_cache import JvmDependencyCache
 from twitter.pants.tasks.nailgun_task import NailgunTask
+from twitter.pants.tasks.zinc_utils import ZincUtils
 
-# TODO: This class is getting huge. Should probably break it up.
 
 # Well known metadata file required to register scalac plugins with nsc.
 _PLUGIN_INFO_FILE = 'scalac-plugin.xml'
@@ -76,7 +71,11 @@ class ScalaCompile(NailgunTask):
 
   def __init__(self, context, workdir=None):
     NailgunTask.__init__(self, context, workdir=context.config.get('scala-compile', 'nailgun_dir'))
-    self._pants_home = get_buildroot()
+
+    color = context.options.scala_compile_color or \
+            context.config.getbool('scala-compile', 'color', default=True)
+
+    self._zinc_utils = ZincUtils(context=context, java_runner=self.runjava, color=color)
 
     self._partition_size_hint = \
       context.options.scala_compile_partition_size_hint \
@@ -89,62 +88,19 @@ class ScalaCompile(NailgunTask):
     if self.check_missing_deps:
       JvmDependencyCache.init_product_requirements(self)
 
-    # We use the scala_compile_color flag if it is explicitly set on the command line.
-    self._color = \
-      context.options.scala_compile_color if context.options.scala_compile_color is not None else \
-      context.config.getbool('scala-compile', 'color', default=True)
-
-    self._compile_profile = context.config.get('scala-compile', 'compile-profile')  # The target scala version.
-    self._zinc_profile = context.config.get('scala-compile', 'zinc-profile')
-    plugins_profile = context.config.get('scala-compile', 'scalac-plugins-profile')
     # for dependency analysis, we need to record the list of analysis cache files generated by splitting
     self.generated_caches = set()
 
-    self._zinc_classpath = nailgun_profile_classpath(self, self._zinc_profile)
-    compiler_classpath = nailgun_profile_classpath(self, self._compile_profile)
-    zinc_jars = ScalaCompile.identify_zinc_jars(compiler_classpath, self._zinc_classpath)
-    self._zinc_jar_args = []
-    for (name, jarpath) in zinc_jars.items():  # The zinc jar names are also the flag names.
-      self._zinc_jar_args.extend(['-%s' % name, jarpath])
-
-    self._plugin_jars = nailgun_profile_classpath(self, plugins_profile) if plugins_profile else []
-
-    self._workdir = context.config.get('scala-compile', 'workdir') if workdir is None else workdir
-    self._classes_dir = os.path.join(self._workdir, 'classes')
-    self._analysis_cache_dir = os.path.join(self._workdir, 'analysis_cache')
-    self._resources_dir = os.path.join(self._workdir, 'resources')
-
-    self._main = context.config.get('scala-compile', 'main')
-
-    self._args = context.config.getlist('scala-compile', 'args')
-    self._jvm_args = context.config.getlist('scala-compile', 'jvm_args')
-    if context.options.scala_compile_warnings:
-      self._args.extend(context.config.getlist('scala-compile', 'warning_args'))
-    else:
-      self._args.extend(context.config.getlist('scala-compile', 'no_warning_args'))
-
-    # Allow multiple flags and also comma-separated values in a single flag.
-    plugin_names = [p for val in context.options.plugins for p in val.split(',')] \
-      if context.options.plugins is not None \
-      else context.config.getlist('scala-compile', 'scalac-plugins', default=[])
-
-    plugin_args = context.config.getdict('scala-compile', 'scalac-plugin-args', default={})
-
-    active_plugins = self.find_plugins(plugin_names)
-
-    for name, jar in active_plugins.items():
-      self._args.append('-Xplugin:%s' % jar)
-      for arg in plugin_args.get(name, []):
-        self._args.append('-P:%s:%s' % (name, arg))
+    workdir = context.config.get('scala-compile', 'workdir') if workdir is None else workdir
+    self._classes_dir = os.path.join(workdir, 'classes')
+    self._analysis_cache_dir = os.path.join(workdir, 'analysis_cache')
+    self._resources_dir = os.path.join(workdir, 'resources')
+    self._depfile_dir = os.path.join(workdir, 'depfiles')
 
     self._confs = context.config.getlist('scala-compile', 'confs')
-    self._depfile_dir = os.path.join(self._workdir, 'depfiles')
 
     artifact_cache_spec = context.config.getlist('scala-compile', 'artifact_caches')
     self.setup_artifact_cache(artifact_cache_spec)
-
-    self._java_home = os.path.dirname(find_java_home())
-    self._ivy_home = context.config.get('ivy', 'cache_dir')
 
   def product_type(self):
     return 'classes'
@@ -164,7 +120,7 @@ class ScalaCompile(NailgunTask):
       with self.context.state('classpath', []) as cp:
         for conf in self._confs:
           cp.insert(0, (conf, self._resources_dir))
-          for jar in self._plugin_jars:
+          for jar in self._zinc_utils.plugin_jars():
             cp.insert(0, (conf, jar))
 
       with self.invalidated(scala_targets, invalidate_dependents=True,
@@ -207,15 +163,10 @@ class ScalaCompile(NailgunTask):
       with self.check_artifact_cache(versioned_target_set,
                                      build_artifacts=[output_dir, depfile, portable_analysis_cache]) as in_cache:
         if in_cache:
-          # Localize the portable analysis cache we got from the cache. Work on a tmpfile, for safety.
-          with temporary_file_path() as tmp_analysis_cache:
-            shutil.copy(portable_analysis_cache, tmp_analysis_cache)
-            rebasings = [
-              (ScalaCompile.IVY_HOME_PLACEHOLDER, self._ivy_home),
-              (ScalaCompile.PANTS_HOME_PLACEHOLDER, self._pants_home),
-            ]
-            self.run_zinc_rebase(cache=tmp_analysis_cache, rebasings=rebasings)
-            shutil.copy(tmp_analysis_cache, analysis_cache)
+          if self._zinc_utils.localize_analysis_cache(portable_analysis_cache, analysis_cache):
+            self.context.log.warn('zinc failed to localise analysis file %s. '\
+                                  'Target may require a full rebuild.' % \
+                                  portable_analysis_cache)
         else:
           self.merge_artifact(versioned_target_set)  # Get what we can from previous builds.
           self.context.log.info('Compiling targets %s' % versioned_target_set.targets)
@@ -227,29 +178,12 @@ class ScalaCompile(NailgunTask):
                                     '\n  '.join(str(t) for t in sources_by_target.keys()))
             else:
               classpath = [jar for conf, jar in cp if conf in self._confs]
-              result = self.compile(classpath, sources, output_dir, analysis_cache, upstream_analysis_caches, depfile)
+              result = self._zinc_utils.compile(classpath, sources, output_dir,
+                                                analysis_cache, upstream_analysis_caches, depfile)
               if result != 0:
-                raise TaskError('%s returned %d' % (self._main, result))
-              # Make the just-created analysis cache portable. Work on a tmpfile, for safety.
-              #
-              # NOTE: We can't port references to deps on the Java home. This is because different JVM
-              # implementations on different systems have different structures, and there's not
-              # necessarily a 1-1 mapping between Java jars on different systems. Instead we simply
-              # drop those references from the analysis cache.
-              #
-              # In practice the JVM changes rarely, and it should be fine to require a full rebuild
-              # in those rare cases.
-              #
-              # TODO: Do this only if we're writing to the artifact cache?
-              with temporary_file_path() as tmp_analysis_cache:
-                shutil.copy(analysis_cache, tmp_analysis_cache)
-                rebasings = [
-                  (self._java_home, ''),  # Erase java deps.
-                  (self._ivy_home, ScalaCompile.IVY_HOME_PLACEHOLDER),
-                  (self._pants_home, ScalaCompile.PANTS_HOME_PLACEHOLDER),
-                ]
-                self.run_zinc_rebase(cache=tmp_analysis_cache, rebasings=rebasings)
-                shutil.copy(tmp_analysis_cache, portable_analysis_cache)
+                raise TaskError('Compile returned %d' % result)
+              if self._zinc_utils.relativize_analysis_cache(analysis_cache, portable_analysis_cache):
+                raise TaskError('Zinc failed to relativize analysis file: %s' % analysis_cache)
     self.post_process(versioned_target_set, upstream_analysis_caches, split_artifact=True)
 
   # Post-processing steps that must happen even for valid targets.
@@ -276,7 +210,7 @@ class ScalaCompile(NailgunTask):
           # Create and Map scala plugin info files to the owning targets.
           for target in vt.targets:
             if is_scalac_plugin(target) and target.classname:
-              basedir = self.write_plugin_info(target)
+              basedir = self._zinc_utils.write_plugin_info(self._resources_dir, target)
               genmap.add(target, basedir, [_PLUGIN_INFO_FILE])
 
     # Update the upstream analysis map.
@@ -308,42 +242,6 @@ class ScalaCompile(NailgunTask):
       collect_sources(target)
     return sources
 
-  def compile(self, classpath, sources, output_dir, analysis_cache, upstream_analysis_caches, depfile):
-    # To pass options to scalac simply prefix with -S.
-    args = ['-S' + x for x in self._args]
-
-    def analysis_cache_full_path(analysis_cache_product):
-      # We expect the argument to be { analysis_cache_dir, [ analysis_cache_file ]}.
-      if len(analysis_cache_product) != 1:
-        raise TaskError('There can only be one analysis cache file per output directory')
-      analysis_cache_dir, analysis_cache_files = analysis_cache_product.iteritems().next()
-      if len(analysis_cache_files) != 1:
-        raise TaskError('There can only be one analysis cache file per output directory')
-      return os.path.join(analysis_cache_dir, analysis_cache_files[0])
-
-    # Strings of <output dir>:<full path to analysis cache file for the classes in that dir>.
-    analysis_map = \
-      OrderedDict([ (k, analysis_cache_full_path(v)) for k, v in upstream_analysis_caches.itermappings() ])
-
-    if len(analysis_map) > 0:
-      args.extend([ '-analysis-map', ','.join(['%s:%s' % kv for kv in analysis_map.items()]) ])
-
-    args.extend([
-      '-analysis-cache', analysis_cache,
-      '-classpath', ':'.join(self._zinc_classpath + classpath),
-      '-output-products', depfile,
-      '-mirror-analysis',
-      '-d', output_dir
-    ])
-
-    if not self._color:
-      args.append('-no-color')
-
-    args.extend(sources)
-
-    self.context.log.debug('Executing: %s %s' % (self._main, ' '.join(args)))
-    return self.run_zinc(args)
-
   # Splits an artifact representing several targets into target-by-target artifacts.
   # Creates an output classes dir, a depfile and an analysis file for each target.
   # Note that it's not OK to create incomplete artifacts here: this is run *after* a zinc invocation,
@@ -368,8 +266,7 @@ class ScalaCompile(NailgunTask):
       for source, classes in classes_by_source.items():
         src = os.path.join(target.target_base, source)
         dst_deps.add(src, classes)
-        source_abspath = os.path.join(self._pants_home, target.target_base, source)
-        sources.append(source_abspath)
+        sources.append(os.path.join(target.target_base, source))
         for cls in classes:
           # Copy the class file.
           dst = os.path.join(dst_output_dir, cls)
@@ -380,7 +277,7 @@ class ScalaCompile(NailgunTask):
       self.generated_caches.add(os.path.join(dst_output_dir, dst_analysis_cache))
     # Split the analysis files.
     if os.path.exists(src_analysis_cache):
-      if self.run_zinc_split(src_cache=src_analysis_cache, splits=[(x[0], x[2]) for x in analysis_splits]):
+      if self._zinc_utils.run_zinc_split(src_cache=src_analysis_cache, splits=[(x[0], x[2]) for x in analysis_splits]):
         raise TaskError, 'zinc failed to split analysis files %s from %s' %\
                          (':'.join([x[2] for x in analysis_splits]), src_analysis_cache)
 
@@ -390,7 +287,7 @@ class ScalaCompile(NailgunTask):
         dst_output_dir = split[1]
         if os.path.exists(dst_analysis_cache):
           rebasings = [(src_output_dir, dst_output_dir)]
-          if self.run_zinc_rebase(cache=dst_analysis_cache, rebasings=rebasings):
+          if self._zinc_utils.run_zinc_rebase(cache=dst_analysis_cache, rebasings=rebasings):
             raise TaskError, 'In split_artifact: zinc failed to rebase analysis file %s' % dst_analysis_cache
 
   # Merges artifacts representing the individual targets in a VersionedTargetSet into one artifact for that set.
@@ -438,132 +335,15 @@ class ScalaCompile(NailgunTask):
               os.path.join(tmpdir, os.path.relpath(src_analysis_cache, self._analysis_cache_dir))
             shutil.copyfile(src_analysis_cache, src_analysis_cache_tmp)
             src_analysis_caches.append(src_analysis_cache_tmp)
-            if self.run_zinc_rebase(cache=src_analysis_cache_tmp, rebasings=[(src_output_dir, dst_output_dir)]):
+            if self._zinc_utils.run_zinc_rebase(cache=src_analysis_cache_tmp, rebasings=[(src_output_dir, dst_output_dir)]):
               self.context.log.warn('In merge_artifact: zinc failed to rebase analysis file %s. ' \
-              'Target may require a full rebuild.' % src_analysis_cache_tmp)
+                                    'Target may require a full rebuild.' % \
+                                    src_analysis_cache_tmp)
 
       dst_deps.save(dst_depfile)
 
-      if self.run_zinc_merge(src_caches=src_analysis_caches, dst_cache=dst_analysis_cache):
-        raise TaskError, 'zinc failed to merge analysis files %s to %s' % \
-                         (':'.join(src_analysis_caches), dst_analysis_cache)
-
-
-  def run_zinc(self, args):
-    zinc_args = [
-      '-log-level', self.context.options.log_level or 'info',
-      ]
-    zinc_args.extend(self._zinc_jar_args)
-    zinc_args.extend(args)
-    return self.runjava(self._main, classpath=self._zinc_classpath, args=zinc_args, jvmargs=self._jvm_args)
-
-  # Run zinc in analysis manipulation mode.
-  def run_zinc_analysis(self, cache, args):
-    zinc_analysis_args = [
-      '-analysis',
-      '-cache', cache,
-      ]
-    zinc_analysis_args.extend(args)
-    return self.run_zinc(args=zinc_analysis_args)
-
-  # src_cache - split this analysis cache.
-  # splits - a list of (sources, dst_cache), where sources is a list of the sources whose analysis
-  #          should be split into dst_cache.
-  def run_zinc_split(self, src_cache, splits):
-    zinc_split_args = [
-      '-split', ','.join(['{%s}:%s' % (':'.join(x[0]), x[1]) for x in splits]),
-      ]
-    return self.run_zinc_analysis(cache=src_cache, args=zinc_split_args)
-
-  # src_caches - a list of caches to merge into dst_cache.
-  def run_zinc_merge(self, src_caches, dst_cache):
-    zinc_merge_args = [
-      '-merge', ':'.join(src_caches),
-      ]
-    return self.run_zinc_analysis(cache=dst_cache, args=zinc_merge_args)
-
-  # cache - the analysis cache to rebase.
-  # rebasings - a list of pairs (rebase_from, rebase_to). Behavior is undefined if any rebase_from
-  # is a prefix of any other, as there is no guarantee that rebasings are applied in a particular order.
-  def run_zinc_rebase(self, cache, rebasings):
-    zinc_rebase_args = [
-      '-rebase', ','.join(['%s:%s' % rebasing for rebasing in rebasings]),
-      ]
-    return self.run_zinc_analysis(cache=cache, args=zinc_rebase_args)
-
-  IVY_HOME_PLACEHOLDER = '/IVY_HOME_PLACEHOLDER'
-  PANTS_HOME_PLACEHOLDER = '/PANTS_HOME_PLACEHOLDER'
-
-  def write_plugin_info(self, target):
-    basedir = os.path.join(self._resources_dir, target.id)
-    with safe_open(os.path.join(basedir, _PLUGIN_INFO_FILE), 'w') as f:
-      f.write(textwrap.dedent('''
-        <plugin>
-          <name>%s</name>
-          <classname>%s</classname>
-        </plugin>
-      ''' % (target.plugin, target.classname)).strip())
-    return basedir
-
-  # These are the names of the various jars zinc needs. They are, conveniently and non-coincidentally,
-  # the names of the flags used to pass the jar locations to zinc.
-  compiler_jar_names = [ 'scala-library', 'scala-compiler' ]  # Compiler version.
-  zinc_jar_names = [ 'compiler-interface', 'sbt-interface' ]  # Other jars zinc needs to be pointed to.
-
-  @staticmethod
-  def identify_zinc_jars(compiler_classpath, zinc_classpath):
-    """Find the named jars in the compiler and zinc classpaths.
-
-    TODO: When profiles migrate to regular pants jar() deps instead of ivy.xml files we can make these
-          mappings explicit instead of deriving them by jar name heuristics.
-    """
-    ret = OrderedDict()
-    ret.update(ScalaCompile.identify_jars(ScalaCompile.compiler_jar_names, compiler_classpath))
-    ret.update(ScalaCompile.identify_jars(ScalaCompile.zinc_jar_names, zinc_classpath))
-    return ret
-
-  @staticmethod
-  def identify_jars(names, jars):
-    jars_by_name = {}
-    jars_and_filenames = [(x, os.path.basename(x)) for x in jars]
-
-    for name in names:
-      jar_for_name = None
-      for jar, filename in jars_and_filenames:
-        if filename.startswith(name):
-          jar_for_name = jar
-          break
-      if jar_for_name is None:
-        raise TaskError('Couldn\'t find jar named %s' % name)
-      else:
-        jars_by_name[name] = jar_for_name
-    return jars_by_name
-
-  def find_plugins(self, plugin_names):
-    """Returns a map from plugin name to plugin jar."""
-    plugin_names = set(plugin_names)
-    plugins = {}
-    # plugin_jars is the universe of all possible plugins and their transitive deps.
-    # Here we select the ones to actually use.
-    for jar in self._plugin_jars:
-      with open_jar(jar, 'r') as jarfile:
-        try:
-          with closing(jarfile.open(_PLUGIN_INFO_FILE, 'r')) as plugin_info_file:
-            plugin_info = ElementTree.parse(plugin_info_file).getroot()
-          if plugin_info.tag != 'plugin':
-            raise TaskError, 'File %s in %s is not a valid scalac plugin descriptor' % (_PLUGIN_INFO_FILE, jar)
-          name = plugin_info.find('name').text
-          if name in plugin_names:
-            if name in plugins:
-              raise TaskError, 'Plugin %s defined in %s and in %s' % (name, plugins[name], jar)
-            # It's important to use relative paths, as the compiler flags get embedded in the zinc
-            # analysis file, and we port those between systems via the artifact cache.
-            plugins[name] = os.path.relpath(jar, self._pants_home)
-        except KeyError:
-          pass
-
-    unresolved_plugins = plugin_names - set(plugins.keys())
-    if len(unresolved_plugins) > 0:
-      raise TaskError, 'Could not find requested plugins: %s' % list(unresolved_plugins)
-    return plugins
+      if self._zinc_utils.run_zinc_merge(src_caches=src_analysis_caches, dst_cache=dst_analysis_cache):
+        self.context.log.warn('zinc failed to merge analysis files %s to %s. ' \
+                              'Target may require a full rebuild.' % \
+                              (':'.join(src_analysis_caches), dst_analysis_cache))
