+++ src/python/twitter/pants/tasks/zinc_analysis_cache.py
@@ -16,11 +16,7 @@
 
 __author__ = 'Mark C. Chu-Carroll'
 
-import os
 from collections import defaultdict
-from twitter.pants.targets.jar_dependency import JarDependency
-from twitter.pants.targets.jvm_target import JvmTarget
-from twitter.pants.tasks import TaskError
 
 
 class ZincMergedAnalysisCache(object):
